requirements-dev.txt
black
build
flake8
pytest
pytest-django
twine
-e .
